Moved the loader’s schema concerns into a dedicated submodule under [packages/web/src/viewer/loader/schema/index.ts](/Users/hal/Documents/VSCodeProjects/n-dx-internal/packages/web/src/viewer/loader/schema/index.ts), with compatibility logic in [compat.ts](/Users/hal/Documents/VSCodeProjects/n-dx-internal/packages/web/src/viewer/loader/schema/compat.ts) and validation in [validate.ts](/Users/hal/Documents/VSCodeProjects/n-dx-internal/packages/web/src/viewer/loader/schema/validate.ts).
