+++ .rex/prd.json
@@ -3553,20 +3553,24 @@
         {
           "id": "15fc8ef8-ac04-4066-b712-9e189d803955",
           "title": "Fix structural in viewer-data-loader (1 finding)",
-          "status": "pending",
+          "status": "completed",
           "level": "feature",
           "description": "- Cohesion (0.4) and coupling (0.6) both sit at warning thresholds; the zone mixes a hook, a service, a utility, and a schema validator — consider whether schema-compat and validate belong in a dedicated schema sub-module to raise cohesion.",
           "priority": "high",
           "source": "sourcevision",
+          "startedAt": "2026-04-20T16:25:47.126Z",
+          "completedAt": "2026-04-20T16:25:47.126Z",
           "children": [
             {
               "id": "506ef736-4331-4f59-8a72-84c6ca8fdd56",
               "title": "Fix structural in viewer-data-loader: Cohesion (0.4) and coupling (0.6) both sit at warning thresholds; the zone mixes",
-              "status": "pending",
+              "status": "completed",
               "level": "task",
               "description": "- Cohesion (0.4) and coupling (0.6) both sit at warning thresholds; the zone mixes a hook, a service, a utility, and a schema validator — consider whether schema-compat and validate belong in a dedicated schema sub-module to raise cohesion.",
               "priority": "high",
               "source": "sourcevision",
+              "startedAt": "2026-04-20T16:21:31.001Z",
+              "completedAt": "2026-04-20T16:25:47.072Z",
               "recommendationMeta": {
                 "findingHashes": [
                   "37c75281a094"
