+++ packages/web/src/viewer/loader/schema/compat.ts
@@ -3,9 +3,10 @@
  * Ensures older output formats can be rendered by newer viewers.
  */
 
+import type { LoadedData } from "../../types.js";
+
 /** Current schema version. Inlined to avoid runtime import from schema/. */
 const SCHEMA_VERSION = "1.0.0";
-import type { LoadedData } from "../types.js";
 
 type ModuleKey = keyof LoadedData;
