+++ packages/web/src/viewer/loader/schema/index.ts
@@ -1,0 +1,9 @@
+export {
+  validateManifest,
+  validateInventory,
+  validateImports,
+  validateZones,
+  validateComponents,
+  validateCallGraph,
+} from "./validate.js";
+export { migrateData, registerMigration } from "./compat.js";
