+++ packages/web/tests/unit/viewer/viewer-data-loader-boundary.test.ts
@@ -1,8 +1,9 @@
 import { describe, it, expect } from "vitest";
-import { readFileSync, readdirSync } from "node:fs";
+import { existsSync, readFileSync, readdirSync } from "node:fs";
 import { join } from "node:path";
 
 const VIEWER_DIR = join(import.meta.dirname!, "..", "..", "..", "src", "viewer");
+const LOADER_DIR = join(VIEWER_DIR, "loader");
 
 function collectTsFiles(dir: string): string[] {
   const files: string[] = [];
@@ -27,5 +28,19 @@ describe("viewer-data-loader boundaries", () => {
     expect(
       readFileSync(join(VIEWER_DIR, "components", "prd-tree", "progressive-loader.ts"), "utf-8"),
     ).toContain("Progressive tree loading for large PRD datasets");
+  });
+
+  it("keeps loader schema helpers inside the loader zone", () => {
+    expect(existsSync(join(VIEWER_DIR, "validate.ts"))).toBe(false);
+
+    expect(readFileSync(join(LOADER_DIR, "data-loader.ts"), "utf-8")).toContain(
+      'from "./schema/index.js"',
+    );
+    expect(readFileSync(join(LOADER_DIR, "schema", "index.ts"), "utf-8")).toContain(
+      'from "./validate.js"',
+    );
+    expect(readFileSync(join(LOADER_DIR, "schema", "index.ts"), "utf-8")).toContain(
+      'from "./compat.js"',
+    );
   });
 });
